inference/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
torch>=2.0.0
xgboost>=2.0.0
pymongo[srv]>=4.5.0
motor>=3.3.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
python-dotenv>=1.0.0
pydantic>=2.0.0
aiokafka>=0.10.0
confluent-kafka>=2.5.0
stream-chat>=4.17.0
prometheus-client>=0.20.0
redis>=5.0.0
py_ecc>=7.0.0
web3>=7.0.0
PyJWT>=2.8.0
